requirements.txt
streamlit>=1.28.0
groq>=0.4.0
Pillow>=10.0.0
